MouseMovementTesting/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

ipython==8.12.3
matplotlib==3.8.2
minimagen==0.0.3
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
pygame==2.5.2
scikit-learn==1.4.0
tensorflow-intel==2.15.0
torch==2.2.0
torchvision==0.17.0
